+++ ctgan/synthesizers/__init__.py
@@ -1,0 +1,16 @@
+from ctgan.synthesizers.base import BaseSynthesizer
+from ctgan.synthesizers.ctgan import CTGANSynthesizer
+from ctgan.synthesizers.tvae import TVAESynthesizer
+
+__all__ = (
+    'BaseSynthesizer',
+    'CTGANSynthesizer',
+    'TVAESynthesizer'
+)
+
+
+def get_all_synthesizers():
+    return {
+        name: globals()[name]
+        for name in __all__
+    }
