+++ ctgan/synthesizers/base.py
@@ -5,7 +5,8 @@ LOGGER = logging.getLogger(__name__)
 
 class BaseSynthesizer:
     """Base class for all default synthesizers of ``CTGAN``.
-    
+
     This should contain the save/load methods.
     """
+    
     pass
